backend/server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stylegenie"
version = "0.1.0"
dependencies = [
    "python-dotenv==1.1.0",
    "fastapi<0.114.0,>=0.113.0",
    "pydantic<3.0.0,>=2.7.0",
    "uvicorn==0.34.0",
    "python-multipart==0.0.20",
    "openai==1.76.0",
    "litellm==1.67.5",
    "fastapi-sessions==0.3.2"
]
